Create "managed" folder
