+++ cli/npm/managed/common.rs
