+++ cli/npm/managed/global.rs
@@ -20,13 +20,13 @@ use deno_runtime::deno_fs::FileSystem;
 use deno_runtime::deno_node::NodePermissions;
 use deno_runtime::deno_node::NodeResolutionMode;
 
-use crate::npm::resolution::NpmResolution;
-use crate::npm::resolvers::common::cache_packages;
 use crate::npm::NpmCache;
 
+use super::common::cache_packages;
 use super::common::types_package_name;
 use super::common::NpmPackageFsResolver;
 use super::common::RegistryReadPermissionChecker;
+use super::resolution::NpmResolution;
 
 /// Resolves packages from the global npm cache.
 #[derive(Debug)]
