+++ cli/npm/managed/local.rs
@@ -42,7 +42,6 @@ use serde::Deserialize;
 use serde::Serialize;
 
 use crate::npm::cache::mixed_case_package_name_encode;
-use crate::npm::resolution::NpmResolution;
 use crate::npm::NpmCache;
 use crate::util::fs::copy_dir_recursive;
 use crate::util::fs::hard_link_dir_recursive;
@@ -50,6 +49,7 @@ use crate::util::fs::hard_link_dir_recursive;
 use super::common::types_package_name;
 use super::common::NpmPackageFsResolver;
 use super::common::RegistryReadPermissionChecker;
+use super::resolution::NpmResolution;
 
 /// Resolver that creates a local node_modules directory
 /// and resolves packages from it.
